backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
pydantic
python-multipart
pypdf
chromadb
python-dotenv
sqlmodel
python-jose[cryptography]
passlib[bcrypt]

